+++ src/sview/part_info.c
@@ -26,7 +26,7 @@
  *
  *  You should have received a copy of the GNU General Public License along
  *  with SLURM; if not, write to the Free Software Foundation, Inc.,
- *  59 Temple Place, Suite 330, Boston, MA  02111-1307  USA.
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA.
 \*****************************************************************************/
 
 #include "src/sview/sview.h"
@@ -51,7 +51,7 @@ typedef struct {
 	uint32_t node_error_cnt;
 	uint32_t node_idle_cnt;
 	List node_ptr_list;
-	uint16_t node_state;
+	uint32_t node_state;
 	partition_info_t* part_ptr;
 	char *reason;
 } sview_part_sub_t;
